refactor: extract ansi styling helpers into standalone module
