+++ src/main.rs
@@ -1,3 +1,4 @@
+mod ansi;
 mod config;
 mod gui;
 mod keychain;
